@@ -437,13 +437,9 @@
 
             <!-- USERNAME -->
             <div class="input-group-custom">
-                <label for="User_Id">User</label>
-              
-                <input id="User_Id"
-                       name="User_Id"
-                       type="text"
-                       placeholder="User ID"
-                       required />
+                <label for="Username">User</label>
+                <!-- If you later add a view model, you can swap this for: asp-for="User_Id" -->
+                <input id="Username" name="Username" type="text" placeholder="Username" required />
             </div>
 
             <!-- PASSWORD -->
