@@ -1,4 +1,4 @@
-﻿@{
+﻿﻿@{
     ViewData["Title"] = "Home";
 }
 <!--https://localhost:7092/-->
